moving timeseries STL function under analysis
